# [client] Improve userspace filter logging performance