## Describe your changes

This improves logging performance when logs are off by eliminating the variadic function call.

  | Test Case      | Variadic         |  LogN   | Improvement |
  |-------------------|-----------------------|-----------------------------|-------------|
  | 2 Args - Disabled | 72.75ns, 32B, 1 alloc | 0.85ns, 0B, 0 allocs        | 86x faster  |
  | 2 Args - Enabled  | 102.2ns, 35B, 1 alloc | 23.45ns, 0B, 0 allocs       | 4.4x faster |
  | 6 Args - Disabled | 112.5ns, 96B, 1 alloc | 0.85ns, 0B, 0 allocs      | 132x faster |
  | 6 Args - Enabled  | 124.9ns, 97B, 1 alloc | 38ns, 0B, 0 allocs        | 3.3x faster |


## Issue ticket number and link

## Stack

<!-- branch-stack -->

### Checklist
- [ ] Is it a bug fix
- [ ] Is a typo/documentation fix
- [ ] Is a feature enhancement
- [ ] It is a refactor
- [ ] Created tests that fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

> By submitting this pull request, you confirm that you have read and agree to the terms of the [Contributor License Agreement](https://github.com/netbirdio/netbird/blob/main/CONTRIBUTOR_LICENSE_AGREEMENT.md).